requirements.txt
black
jinja2
jsonschema
PySDL2
py-ubjson
pygame
pylint
